pyproject.toml
[project]
name = "raspipool"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "homeassistant",
    "pyserial",
    "strands-agents",
    "ollama==0.5.1"
]

[tool.ruff.lint]
ignore = [
    "F841",
    "E701"
]
